# api-spec/openapi.yaml
openapi: 3.0.3
info:
  title: Fintech SaaS API
  version: 1.0.0
  description: API de gestão financeira multi-tenant.
servers:
  - url: http://localhost:8080
    description: Desenvolvimento local
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    TransactionType:
      type: string
      enum: [INCOME, EXPENSE]
...
    BudgetCycleStatus:
      type: string
      enum: [OPEN, CLOSED]
    BudgetItemSource:
      type: string
      enum: [MANUAL, RECURRING, INSTALLMENT]
    BudgetItemStatus:
      type: string
      enum: [PENDING, REALIZED, SKIPPED]
    BudgetCycleOpenRequest:
      type: object
      required: [referenceMonth, startDay]
      properties:
        referenceMonth:
          type: string
          pattern: '^\d{4}-\d{2}$'
          example: "2026-06"
        startDay:
          type: integer
          minimum: 1
          maximum: 28
          example: 1
          description: "Dia do mês em que o ciclo começa (1 = mês calendário)"
    BudgetCycleSummary:
      type: object
      properties:
        openingBalance:
          type: number
          format: double
        plannedIncome:
          type: number
          format: double
        plannedExpense:
          type: number
          format: double
        projectedBalance:
          type: number
          format: double
        realizedIncome:
          type: number
          format: double
        realizedExpense:
          type: number
          format: double
        unplannedExpenses:
          type: number
          format: double
        availableToSpend:
          type: number
          format: double
        dailyAllowance:
          type: number
          format: double
        remainingDays:
          type: integer
        pendingCount:
          type: integer
    BudgetItemResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        description:
          type: string
        amount:
          type: number
          format: double
        type:
          $ref: '#/components/schemas/TransactionType'
        expectedDate:
          type: string
          format: date
        source:
          $ref: '#/components/schemas/BudgetItemSource'
        status:
          $ref: '#/components/schemas/BudgetItemStatus'
        categoryId:
          type: string
          format: uuid
          nullable: true
        categoryName:
          type: string
          nullable: true
        accountId:
          type: string
          format: uuid
          nullable: true
        accountName:
          type: string
          nullable: true
        transactionId:
          type: string
          format: uuid
          nullable: true
        installmentGroupId:
          type: string
          format: uuid
          nullable: true
    BudgetCycleResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        referenceMonth:
          type: string
          example: "2026-06"
          description: "Mês de referência do ciclo (yyyy-MM)"
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        openingBalance:
          type: number
          format: double
        status:
          $ref: '#/components/schemas/BudgetCycleStatus'
        summary:
          $ref: '#/components/schemas/BudgetCycleSummary'
        items:
          type: array
          items:
            $ref: '#/components/schemas/BudgetItemResponse'
    BudgetItemCreateRequest:
      type: object
      required: [description, amount, type, expectedDate]
      properties:
        description:
          type: string
        amount:
          type: number
          format: double
        type:
          $ref: '#/components/schemas/TransactionType'
        expectedDate:
          type: string
          format: date
        categoryId:
          type: string
          format: uuid
          nullable: true
        accountId:
          type: string
          format: uuid
          nullable: true
...
    BudgetCyclePageResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/BudgetCycleResponse'
        totalElements:
          type: integer
        totalPages:
          type: integer
        number:
          type: integer
paths:
...
  /api/budget-cycles:
    get:
      tags: [budget]
      operationId: listBudgetCycles
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 12
      responses:
        '200':
          description: Histórico de ciclos paginado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BudgetCyclePageResponse'
    post:
      tags: [budget]
      operationId: openBudgetCycle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BudgetCycleOpenRequest'
      responses:
        '201':
          description: Ciclo aberto
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BudgetCycleResponse'
        '409':
          description: Ciclo já existe ou período conflita
  /api/budget-cycles/current:
    get:
      tags: [budget]
      operationId: getCurrentBudgetCycle
      responses:
        '200':
          description: Ciclo OPEN atual com itens
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BudgetCycleResponse'
        '404':
          description: Nenhum ciclo aberto
  /api/budget-cycles/{id}:
    get:
      tags: [budget]
      operationId: getBudgetCycle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Ciclo com itens
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BudgetCycleResponse'
  /api/budget-cycles/{id}/close:
    post:
      tags: [budget]
      operationId: closeBudgetCycle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Ciclo fechado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BudgetCycleResponse'
  /api/budget-cycles/{id}/sync-installments:
    post:
      tags: [budget]
      operationId: syncInstallments
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Parcelas sincronizadas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BudgetCycleResponse'
  /api/budget-cycles/{cycleId}/items:
    get:
      tags: [budget]
      operationId: listBudgetItems
      parameters:
        - name: cycleId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Lista de itens do ciclo
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BudgetItemResponse'
    post:
      tags: [budget]
      operationId: createBudgetItem
      parameters:
        - name: cycleId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BudgetItemCreateRequest'
      responses:
        '201':
          description: Item criado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BudgetItemResponse'
...