# backend/src/main/java/com/fintech/api/dto/budget/BudgetCycleResponseDTO.java
package com.fintech.api.dto.budget;

import com.fintech.api.domain.budget.BudgetCycle;
import com.fintech.api.domain.budget.BudgetItem;
import com.fintech.api.domain.enums.BudgetCycleStatus;
import java.math.BigDecimal;
import java.time.LocalDate;
import java.util.List;
import java.util.UUID;

public record BudgetCycleResponseDTO(
    UUID id,
    String referenceMonth,
    LocalDate startDate,
    LocalDate endDate,
    BigDecimal openingBalance,
    BudgetCycleStatus status,
    BudgetCycleSummaryDTO summary,
    List<BudgetItemResponseDTO> items
) {
    public static BudgetCycleResponseDTO fromEntity(BudgetCycle cycle, List<BudgetItem> items, BudgetCycleSummaryDTO summary) {
        List<BudgetItemResponseDTO> itemDTOs = items.stream()
            .map(BudgetItemResponseDTO::fromEntity)
            .toList();
        return new BudgetCycleResponseDTO(
            cycle.getId(),
            cycle.getReferenceMonth(),
            cycle.getStartDate(),
            cycle.getEndDate(),
            cycle.getOpeningBalance(),
            cycle.getStatus(),
            summary,
            itemDTOs
        );
    }
}